+++ index.html
@@ -35,7 +35,7 @@
 
             <!-- Изначальное состояние поля для загрузки изображения -->
             <fieldset class="img-upload__start">
-              <input type="file" id="upload-file" class="img-upload__input  visually-hidden" name="filename" required>
+              <input type="file"accept='image/*' id="upload-file" class="img-upload__input  visually-hidden" name="filename" required>
               <label for="upload-file" class="img-upload__label  img-upload__control">Загрузить</label>
             </fieldset>
 
@@ -236,6 +236,5 @@
   </template>
   <script src="vendor/pristine/pristine.min.js"></script>
   <script type="module" src="js/main.js"></script>
-  <script type="module" src="js/image-loader.js"></script>
 </body>
 </html>
