+++ js/edit-popup.js
@@ -1,0 +1,81 @@
+import { isEscapeKey } from './utils.js';
+
+const MAX_COUNT_HASHTAG = 5;
+const HASTAG_REGEX = /^#[a-zа-яё0-9]{1,19}$/i;
+
+const documentBody = document.querySelector('body');
+const form = documentBody.querySelector('.img-upload__form');
+const editForm = form.querySelector('.img-upload__overlay');
+const closeButton = form.querySelector('.img-upload__cancel');
+const imageLoadingField = form.querySelector('.img-upload__input ');
+const hashtagField = form.querySelector('.text__hashtags');
+const commentField = form.querySelector('.text__description');
+
+const pristine = new Pristine (form, {
+  classTo: 'img-upload__field-wrapper',
+  errorTextParent: 'img-upload__field-wrapper',
+  errorTextClass: 'img-upload__field-wrapper-error-text',
+});
+
+const showForm = () => {
+  editForm.classList.remove('hidden');
+  documentBody.classList.add('modal-open');
+  document.addEventListener('keydown', onDocumentKeydown);
+  closeButton.addEventListener('click', onCloseButtonClick);
+  initValidate();
+};
+
+const hideForm = () => {
+  editForm.classList.add('hidden');
+  documentBody.classList.remove('modal-open');
+  document.removeEventListener('keydown', onDocumentKeydown);
+  closeButton.removeEventListener('click', onCloseButtonClick);
+  imageLoadingField.value = '';
+  hashtagField.value = '';
+  commentField.value = '';
+  pristine.reset();
+};
+
+const extarctHastag = (value) => value.trim().split(' ').filter((element) => element.length > 0);
+
+const isValidHastag = (value) => extarctHastag(value).every((element) => HASTAG_REGEX.test(element));
+
+const isAmountHastag = (value) => extarctHastag(value).length <= MAX_COUNT_HASHTAG;
+
+const isUniqueHastag = (value) => {
+  const oneCaseHastags = extarctHastag(value).map((element) => element.toLowerCase());
+  return new Set(oneCaseHastags).size === oneCaseHastags.length;
+};
+
+function initValidate () {
+  pristine.addValidator(hashtagField, isAmountHastag, `Нельзя вводить более ${MAX_COUNT_HASHTAG} хештегов :-(`);
+  pristine.addValidator(hashtagField, isValidHastag, 'Хештег невалиден :-(');
+  pristine.addValidator(hashtagField, isUniqueHastag, 'Хештеги не должны повторяться :-(');
+}
+
+function onDocumentKeydown (evt) {
+  if (isEscapeKey(evt) && document.activeElement !== hashtagField
+  && document.activeElement !== commentField) {
+    evt.preventDefault();
+    hideForm();
+  }
+}
+
+const onImageLoadingFieldChange = (evt) => {
+  evt.preventDefault();
+  const selectedFiel = imageLoadingField.files[0];
+  if(selectedFiel.type.startsWith('image/') || /\.(jpg|jpeg|png|gif)$/i.test(selectedFiel.name)){
+    showForm();
+  }
+};
+
+function onCloseButtonClick (evt) {
+  evt.preventDefault();
+  hideForm();
+}
+
+const openEditPopup = () => {
+  imageLoadingField.addEventListener('change', onImageLoadingFieldChange);
+};
+
+export { openEditPopup };
