+++ js/gallery.js
@@ -4,10 +4,7 @@ import { renderThumbnails } from './thumbnail.js';
 const container = document.querySelector('.pictures');
 const renderGallery = (picutres) => {
   container.addEventListener('click', (evt) => {
-    //evt.preventDefault();
-
     const thumbnail = evt.target.closest('[data-thumbnail-id]');
-
     if(!thumbnail) {
 
       return;
